+++ django_resilient_webhook/utilities/short_sha.py
@@ -1,0 +1,6 @@
+import hashlib
+
+
+def short_sha(value, digits=8):
+    """Calculate a short sha from an input string, with a certain degree of uniqueness"""
+    return hashlib.sha256(value.encode("utf-8")).hexdigest()[:digits]
